requirements-pipeline.txt
absl-py==0.14.1
albumentations==1.1.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudpickle==1.3.0
ConfigArgParse==1.5.3
connection-pool==0.0.3
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
datrie==0.8.2
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.3
dm-tree==0.1.6
docutils==0.18.1
entrypoints==0.3
filelock==3.4.1
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==2.10.0
idna==3.3
imageio==2.9.0
immutables==0.16
importlib-metadata==4.8.1
importlib-resources==5.4.0
ipywidgets==7.6.5
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
matplotlib==3.3.4
mistune==0.8.4
nest-asyncio==1.5.1
networkx==2.5.1
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
openslide-python==1.1.2
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.8.0
ptyprocess==0.7.0
PuLP==2.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
qudida==0.0.4
ratelimiter==1.2.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.20.0
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
snakemake==6.15.5
sniffio==1.2.0
stopit==1.1.2
tabulate==0.8.9
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-probability==0.11.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2020.9.3
toposort==1.7
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.13.2
xgboost==1.3.3
zipp==3.6.0
